MAILBOX-364 Move EventDelivery class to event-memory module
