+++ mailbox/event/event-memory/src/main/java/org/apache/james/mailbox/events/delivery/SynchronousEventDelivery.java
@@ -17,36 +17,39 @@
  * under the License.                                           *
  ****************************************************************/
 
-package org.apache.james.mailbox.store.event;
+package org.apache.james.mailbox.events.delivery;
 
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
-import java.util.concurrent.ThreadFactory;
-
-import javax.annotation.PreDestroy;
+import javax.inject.Inject;
 
 import org.apache.james.mailbox.Event;
 import org.apache.james.mailbox.MailboxListener;
-import org.apache.james.util.concurrent.NamedThreadFactory;
+import org.apache.james.metrics.api.MetricFactory;
+import org.apache.james.metrics.api.TimeMetric;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
-public class AsynchronousEventDelivery implements EventDelivery {
+public class SynchronousEventDelivery implements EventDelivery {
 
-    private final ExecutorService threadPoolExecutor;
-    private final SynchronousEventDelivery synchronousEventDelivery;
+    private static final Logger LOGGER = LoggerFactory.getLogger(SynchronousEventDelivery.class);
 
-    public AsynchronousEventDelivery(int threadPoolSize, SynchronousEventDelivery synchronousEventDelivery) {
-        ThreadFactory threadFactory = NamedThreadFactory.withClassName(getClass());
-        this.threadPoolExecutor = Executors.newFixedThreadPool(threadPoolSize, threadFactory);
-        this.synchronousEventDelivery = synchronousEventDelivery;
+    private final MetricFactory metricFactory;
+
+    @Inject
+    public SynchronousEventDelivery(MetricFactory metricFactory) {
+        this.metricFactory = metricFactory;
     }
 
     @Override
     public void deliver(MailboxListener mailboxListener, Event event) {
-        threadPoolExecutor.submit(() -> synchronousEventDelivery.deliver(mailboxListener, event));
-    }
-
-    @PreDestroy
-    public void stop() {
-        threadPoolExecutor.shutdownNow();
+        TimeMetric timer = metricFactory.timer("mailbox-listener-" + mailboxListener.getClass().getSimpleName());
+        try {
+            mailboxListener.event(event);
+        } catch (Throwable throwable) {
+            LOGGER.error("Error while processing listener {} for {}",
+                    mailboxListener.getClass().getCanonicalName(), event.getClass().getCanonicalName(),
+                    throwable);
+        } finally {
+            timer.stopAndPublish();
+        }
     }
 }
