+++ mailbox/pom.xml
@@ -62,6 +62,7 @@
         <module>tools/indexer</module>
         <module>tools/jpa-migrator</module>
         <module>tools/maildir-utils</module>
+        <module>event/event-memory</module>
     </modules>
 
     <issueManagement>
