+++ mailbox/spring/src/main/resources/META-INF/spring/event-system.xml
@@ -28,18 +28,18 @@
         <constructor-arg index="1" ref="event-registry"/>
     </bean>
 
-    <bean id="synchronous-event-delivery" class="org.apache.james.mailbox.store.event.SynchronousEventDelivery" lazy-init="true">
+    <bean id="synchronous-event-delivery" class="org.apache.james.mailbox.events.delivery.SynchronousEventDelivery" lazy-init="true">
         <constructor-arg index="0" ref="metricFactory"/>
     </bean>
 
     <bean id="event-registry" class="org.apache.james.mailbox.store.event.MailboxListenerRegistry"/>
 
-    <bean id="asynchronous-event-delivery" class="org.apache.james.mailbox.store.event.AsynchronousEventDelivery" lazy-init="true">
+    <bean id="asynchronous-event-delivery" class="org.apache.james.mailbox.events.delivery.AsynchronousEventDelivery" lazy-init="true">
         <constructor-arg index="0" ref="${event.delivery.thread.count}"/>
         <constructor-arg index="1" ref="synchronous-event-delivery"/>
     </bean>
 
-    <bean id="mixed-event-delivery" class="org.apache.james.mailbox.store.event.MixedEventDelivery" lazy-init="true">
+    <bean id="mixed-event-delivery" class="org.apache.james.mailbox.events.delivery.MixedEventDelivery" lazy-init="true">
         <constructor-arg index="0" ref="asynchronous-event-delivery"/>
         <constructor-arg index="1" ref="synchronous-event-delivery"/>
     </bean>
